Add log_bin_implicit_delete setting
On a setup with a master, two intermediate masters and a slave like this:
```
           A
         /   \
        B     C
       /
      D
```

If you have a table like this:
CREATE TABLE t1 (id INT PRIMARY KEY) ENGINE=MEMORY;
INSERT INTO t1 VALUES (1);

And then you restart C and do a `SELECT * FROM t1`:
* the MEMORY engine marks the table as implicitly emptied
* the server writes a "DELETE FROM `test`.`t1`" to the binlog
* If GTID is enabled this creates a GTID local to C

If you now try to move D to C this happens:
* If the DELETE is still in the binary log it will fetch it and execute it, this is OK.
* If the DELETE has been purged from the binlog (e.g. after binlog_expire_logs_seconds)
  then replication will fail with ER_MASTER_HAS_PURGED_REQUIRED_GTIDS

This commit does two things:
1. It adds a comment to the DELETE to make it easier to discover where it came from.
2. It adds a log_bin_implicit_delete to allow the administrator to configure the
   server not to write these DELETE's to the binlog.

* https://dev.mysql.com/doc/refman/8.0/en/replication-features-memory.html
* https://bugs.mysql.com/bug.php?id=77729

+++ sql/sql_base.cc
@@ -3790,7 +3790,7 @@ static bool open_table_entry_fini(THD *thd, TABLE_SHARE *share,
   if (unlikely(entry->file->implicit_emptied) &&
       (!thd->lex || !thd->lex->m_IS_table_stats.is_reading_stats_by_open())) {
     entry->file->implicit_emptied = 0;
-    if (mysql_bin_log.is_open()) {
+    if ((mysql_bin_log.is_open()) && (log_bin_implicit_delete)) {
       bool error = false;
       String temp_buf;
       error = temp_buf.append("DELETE FROM ");
@@ -3798,6 +3798,7 @@ static bool open_table_entry_fini(THD *thd, TABLE_SHARE *share,
       error = temp_buf.append(".");
       append_identifier(thd, &temp_buf, share->table_name.str,
                         strlen(share->table_name.str));
+      error = temp_buf.append(" /* added by mysqld because of implicitly emptied table */");
       if (mysql_bin_log.write_dml_directly(thd, temp_buf.c_ptr_safe(),
                                            temp_buf.length()))
         return true;
